\name{CONTundirv7-package}
\alias{CONTundirv7-package}
\alias{CONTundirv7}
\docType{package}
\title{
\packageTitle{CONTundirv7}
}
\description{
\packageDescription{CONTundirv7}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CONTundirv7}
\packageIndices{CONTundirv7}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CONTundirv7}

Maintainer: \packageMaintainer{CONTundirv7}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
